# packages/contracts/openapi/dealdrop.v1.yaml
openapi: 3.1.0
info:
  title: DealDrop API
  version: 2.0.0
  description: >
    Contract-first API foundation for the DealDrop mobile app, admin portal,
    and async worker ecosystem.
servers:
  - url: https://api.dealdrop.app
...
paths:
...
  /v1/listings/{listingId}:
    get:
      tags: [Listings]
      summary: Return a listing detail record.
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/venues/{venueId}:
    get:
      tags: [Listings]
      summary: Return a venue detail record.
      parameters:
        - $ref: '#/components/parameters/VenueId'
      responses:
        '200':
          description: Venue detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/listings/{listingId}/confirm:
    post:
      tags: [Contributions]
      summary: Record a validity confirmation.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                proofAssetKey:
                  type: string
      responses:
        '200':
          description: Updated trust state.
          content:
            application/json:
              schema:
                type: object
                properties:
                  confidenceScore:
                    type: number
                  trustBand:
                    $ref: '#/components/schemas/TrustBand'
  /v1/listings/{listingId}/report-expired:
    post:
      tags: [Contributions]
      summary: Report a listing as expired or incorrect.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportExpired'
      responses:
        '200':
          description: Report accepted.
...
  /v1/auth/sign-in:
    post:
      tags: [Auth]
      summary: Sign in with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /v1/auth/sign-up:
    post:
      tags: [Auth]
      summary: Create a local development account and return the authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/AuthRequest'
                - type: object
                  required: [displayName, homeNeighborhood]
      responses:
        '200':
          description: Authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /v1/auth/bootstrap:
    post:
      tags: [Auth]
      summary: Upsert the Supabase-authenticated user profile and return the app session contract.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthBootstrapRequest'
      responses:
        '200':
          description: Authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /v1/notifications:
    get:
      tags: [Notifications]
      summary: Return the in-app notification inbox.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
  /v1/notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      summary: Mark an inbox notification as read.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated notification record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRecord'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/me/preferences:
    get:
      tags: [Notifications]
      summary: Return the active user's notification preferences.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Preferences payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
    put:
      tags: [Notifications]
      summary: Update the active user's notification preferences.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Updated preferences.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
  /v1/devices/register:
    post:
      tags: [Notifications]
      summary: Register a device for push and inbox targeting.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegistration'
      responses:
        '200':
          description: Registered device details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRegistration'
  /v1/devices/{deviceId}:
    delete:
      tags: [Notifications]
      summary: Unregister a device.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Device unregistered.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ListingId:
      in: path
      name: listingId
      required: true
      schema:
        type: string
    VenueId:
      in: path
      name: venueId
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found.
  schemas:
    TrustBand:
      type: string
      enum: [founder_verified, merchant_confirmed, user_confirmed, recently_updated, needs_recheck, disputed]
...
    TrustSummary:
      type: object
      properties:
        band:
          $ref: '#/components/schemas/TrustBand'
        explanation:
          type: string
        confidenceScore:
          type: number
        freshUntilAt:
          type: string
          format: date-time
        recheckAfterAt:
          type: string
          format: date-time
        proofCount:
          type: integer
        recentConfirmations:
          type: integer
        disputeCount:
          type: integer
    ListingCard:
      type: object
      required:
        [id, venueId, venueName, title, neighborhood, categoryLabel, scheduleLabel, trustBand, freshnessText]
      properties:
        id:
          type: string
        venueId:
          type: string
        venueName:
          type: string
        title:
          type: string
        neighborhood:
          type: string
        categoryLabel:
          type: string
        scheduleLabel:
          type: string
        trustBand:
          $ref: '#/components/schemas/TrustBand'
        freshnessText:
          type: string
        valueNote:
          type: string
        affordabilityLabel:
          type: string
        distanceMiles:
          type: number
        rating:
          type: number
        cuisine:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        confidenceScore:
          type: number
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            type: string
        saved:
          type: boolean
    ListingDetail:
      allOf:
        - $ref: '#/components/schemas/ListingCard'
        - type: object
          properties:
            venueAddress:
              type: string
            description:
              type: string
            conditions:
              type: string
            sourceNote:
              type: string
            offers:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  originalPrice:
                    type: number
                  dealPrice:
                    type: number
                  currency:
                    type: string
            confidenceScore:
              type: number
            freshUntilAt:
              type: string
              format: date-time
            recheckAfterAt:
              type: string
              format: date-time
            proofCount:
              type: integer
            trustSummary:
              $ref: '#/components/schemas/TrustSummary'
    VenueDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        neighborhood:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        rating:
          type: number
        listingIds:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        activeListingCount:
          type: integer
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        homeNeighborhood:
          type: string
        role:
          type: string
          enum: [user, moderator, admin]
        verifiedContributor:
          type: boolean
...
    ReportExpired:
      type: object
      properties:
        reason:
          type: string
        notes:
          type: string
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        displayName:
          type: string
        homeNeighborhood:
          type: string
    AuthBootstrapRequest:
      type: object
      properties:
        displayName:
          type: string
        homeNeighborhood:
          type: string
          default: Midtown
    AuthSession:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [user, moderator, admin]
        verifiedContributor:
          type: boolean
    AuthResponse:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/AuthSession'
        profile:
          $ref: '#/components/schemas/Profile'
    NotificationRecord:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [contribution_resolved, points_finalized, trust_status_changed, listing_reported_stale, moderation_update]
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
        deepLink:
          type: string
          nullable: true
    NotificationsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationRecord'
        unreadCount:
          type: integer
    UserPreferences:
      type: object
      properties:
        contributionResolved:
          type: boolean
        pointsFinalized:
          type: boolean
        trustStatusChanged:
          type: boolean
        marketingAnnouncements:
          type: boolean
    DeviceRegistration:
      type: object
      properties:
        deviceId:
          type: string
        platform:
          type: string
          enum: [ios, android, web, macos, windows, linux]
        pushToken:
          type: string
          nullable: true
        locale:
          type: string
        notificationsEnabled:
          type: boolean
...